Consolidate navigators.
Flatten the project structure, address uses of nav service in widgets.
