+++ app/lib/core/view_models/sign_in_view_model.dart
@@ -1,10 +1,12 @@
-import 'package:app/core/services/authentication/auth_service.dart';
-import 'package:app/core/services/dialog/dialog_service.dart';
-import 'package:app/core/services/navigation/nav_service.dart';
+import 'package:app/core/services/auth_service.dart';
+import 'package:app/core/services/dialog_service.dart';
+import 'package:app/core/services/nav_service.dart';
 import 'package:app/core/services/service_locator.dart';
 import 'package:app/core/view_models/base_view_model.dart';
 import 'package:app/ui/common/view_state.dart';
-import 'package:app/ui/views/home/home_screen.dart';
+import 'package:app/ui/views/home_screen.dart';
+import 'package:app/ui/views/sign_up_screen.dart';
+import 'package:app/ui/views/welcome_screen.dart';
 import 'package:flutter/material.dart';
 
 class SignInViewModel extends BaseViewModel {
@@ -13,7 +15,11 @@ class SignInViewModel extends BaseViewModel {
       locator<AuthenticationService>();
   final NavigationService _navigationService = locator<NavigationService>();
 
-  void navigateToWelcomeScreen() => _navigationService.pop();
+  void navigateToWelcomeScreen() =>
+      _navigationService.navigateBackUntil(WelcomeScreen.route);
+
+  void navigateToSignUpScreen() =>
+      _navigationService.navigateAndReplace(SignUpScreen.route);
 
   Future<void> signIn({
     @required String email,
