+++ app/lib/core/view_models/signup_view_model.dart
@@ -1,10 +1,11 @@
-import 'package:app/core/services/authentication/auth_service.dart';
-import 'package:app/core/services/dialog/dialog_service.dart';
-import 'package:app/core/services/navigation/nav_service.dart';
+import 'package:app/core/services/auth_service.dart';
+import 'package:app/core/services/dialog_service.dart';
+import 'package:app/core/services/nav_service.dart';
 import 'package:app/core/services/service_locator.dart';
 import 'package:app/core/view_models/base_view_model.dart';
 import 'package:app/ui/common/view_state.dart';
-import 'package:app/ui/views/signin/sign_in_screen.dart';
+import 'package:app/ui/views/sign_in_screen.dart';
+import 'package:app/ui/views/welcome_screen.dart';
 import 'package:flutter/material.dart';
 
 class SignUpViewModel extends BaseViewModel {
@@ -13,7 +14,11 @@ class SignUpViewModel extends BaseViewModel {
   final DialogService _dialogService = locator<DialogService>();
   final NavigationService _navigationService = locator<NavigationService>();
 
-  void navigateToWelcomeScreen() => _navigationService.pop();
+  void navigateToWelcomeScreen() =>
+      _navigationService.navigateBackUntil(WelcomeScreen.route);
+
+  void navigateToSignInScreen() =>
+      _navigationService.navigateAndReplace(SignInScreen.route);
 
   Future<void> signUp({
     @required String firstName,
