+++ app/lib/core/view_models/welcome_screen_view_model.dart
@@ -1,10 +1,10 @@
-import 'package:app/core/services/authentication/auth_service.dart';
-import 'package:app/core/services/navigation/nav_service.dart';
+import 'package:app/core/services/auth_service.dart';
+import 'package:app/core/services/nav_service.dart';
 import 'package:app/core/services/service_locator.dart';
 import 'package:app/core/view_models/base_view_model.dart';
-import 'package:app/ui/views/home/home_screen.dart';
-import 'package:app/ui/views/signin/sign_in_screen.dart';
-import 'package:app/ui/views/signup/sign_up_screen.dart';
+import 'package:app/ui/views/home_screen.dart';
+import 'package:app/ui/views/sign_in_screen.dart';
+import 'package:app/ui/views/sign_up_screen.dart';
 
 class WelcomeScreenViewModel extends BaseViewModel {
   final NavigationService _navigationService = locator<NavigationService>();
