+++ app/lib/ui/views/sign_in_screen.dart
@@ -1,16 +1,11 @@
-import 'package:app/core/services/navigation/nav_service.dart';
-import 'package:app/core/services/service_locator.dart';
 import 'package:app/core/view_models/sign_in_view_model.dart';
 import 'package:app/ui/common/view_state.dart';
-import 'package:app/ui/views/signup/sign_up_screen.dart';
-import 'package:app/ui/views/welcome/welcome_screen.dart';
 import 'package:app/ui/widgets/utility/busy_overlay_screen.dart';
 import 'package:app/ui/widgets/utility/template_base_view_model.dart';
 import 'package:flutter/material.dart';
 
 class SignInScreen extends StatefulWidget {
   static const route = '/signIn';
-  final NavigationService _navigationService = locator<NavigationService>();
   final formKey = GlobalKey<FormState>();
 
   SignInScreen({
@@ -124,8 +119,7 @@ class _SignInScreenState extends State<SignInScreen> {
                           ),
                           TextButton(
                             child: Text('Do not have an account? Sign Up Here'),
-                            onPressed: () => widget._navigationService
-                                .navigateTo(SignUpScreen.route),
+                            onPressed: () => model.navigateToSignUpScreen(),
                           ),
                           Padding(
                             padding: EdgeInsets.only(
@@ -134,8 +128,7 @@ class _SignInScreenState extends State<SignInScreen> {
                           ),
                           TextButton(
                             child: Text('Go back to welcome screen'),
-                            onPressed: () => widget._navigationService
-                                .navigateTo(WelcomeScreen.route),
+                            onPressed: () => model.navigateToWelcomeScreen(),
                           ),
                         ],
                       ),
