+++ app/test/main_test.dart
@@ -1,11 +1,11 @@
-import 'package:app/core/services/authentication/auth_service.dart';
-import 'package:app/core/services/dialog/dialog_service.dart';
-import 'package:app/core/services/navigation/nav_service.dart';
+import 'package:app/core/services/auth_service.dart';
+import 'package:app/core/services/dialog_service.dart';
+import 'package:app/core/services/nav_service.dart';
 import 'package:app/core/utilities/optional.dart';
 import 'package:app/core/view_models/welcome_screen_view_model.dart';
 import 'package:app/humane_transport_app.dart';
 import 'package:app/test/mock/test_service_locator.dart';
-import 'package:app/ui/views/welcome/welcome_screen.dart';
+import 'package:app/ui/views/welcome_screen.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
 import 'package:mockito/mockito.dart';
