+++ app/test/ui/views/active/atr_editing_screen_test.dart
@@ -1,6 +1,6 @@
 import 'package:app/core/models/animal_transport_record.dart';
-import 'package:app/core/services/dialog/dialog_service.dart';
-import 'package:app/core/services/navigation/nav_service.dart';
+import 'package:app/core/services/dialog_service.dart';
+import 'package:app/core/services/nav_service.dart';
 import 'package:app/core/view_models/active_screen_view_model.dart';
 import 'package:app/test/test_data.dart';
 import 'package:app/ui/views/active/atr_editing_screen.dart';
